ecc384 driver: Add calls to cfi_assert_eq_n_words.
