+++ FROZEN_IMAGES.sha384sum
@@ -1,3 +1,3 @@
 # WARNING: Do not update this file without the approval of the Caliptra TAC
-8645143a3e1d5638c3b037b49a9e0ea7f192171de8279a27d772870e6e5c4b4e787820c504ca87d03a0b624fb2251cda  caliptra-rom-no-log.bin
-1fb1c9e4cf35688b4089559e847e6948e42f4c994b2e9b5b6dccfa43d8706a9fb14de836ddd3b052f97aa810b3e73280  caliptra-rom-with-log.bin
+2b92c15fb71345d8e6576f597dd0e7ba609cf4769cd8d177f09f3316b9b2a850caa90d67398cf6c7ed2174282287566f  caliptra-rom-no-log.bin
+49b3f504a3db1b597b77cc8c214f17d94e94e2f3d66647a98481989e8497ca3fc5a4705d6bcc9419b076a293215346b4  caliptra-rom-with-log.bin
